Add an explanation regarding the after-release day + tiers
Add an explanation regarding the after-release day + tiers.

Also, remove the part about opening issues when the feature implementation is not finished, this is not true for a long time because we do not close the central issue until everything is completed. Unless we thought the community could do that.